medlingo-backend/speech_to_text_service/requirements.txt
# ---------- common ----------
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.0.0  # <-- NEW
sqlalchemy>=2.0.35
alembic>=1.13.2
python-multipart>=0.0.9,<1.0

# ---------- service-specific ----------
# speech_to_text_service
azure-cognitiveservices-speech>=1.40.0   # or openai, deepgram, etc.

# text_translator_service
azure-ai-translation-text>=1.0.0

# text_to_speech_service
azure-cognitiveservices-speech>=1.40.0   # TTS uses the same SDK

# audit_service
pyodbc>=5.1.0                     # <-- NEW